Tagging v1.4.3

+++ includes/class-schema-scalpel-uninstaller.php
@@ -20,10 +20,15 @@ class Schema_Scalpel_Uninstaller {
 	 */
 	public static function uninstall() {
 		global $wpdb;
-		$schemas = $wpdb->prefix . 'scsc_custom_schemas';
-		$wpdb->query( $wpdb->prepare( 'DROP TABLE IF EXISTS %1s;', $schemas ) );
-		$settings = $wpdb->prefix . 'scsc_settings';
-		$wpdb->query( $wpdb->prepare( 'DROP TABLE IF EXISTS %1s;', $settings ) );
+		$is_delete = $wpdb->get_results( $wpdb->prepare( "SELECT * FROM %1s WHERE setting_key = 'delete_on_uninstall';", $wpdb->prefix . 'scsc_settings' ), ARRAY_A )[0]['setting_value'];
+
+		if ( '1' === $is_delete ) {
+			$schemas = $wpdb->prefix . 'scsc_custom_schemas';
+			$wpdb->query( $wpdb->prepare( 'DROP TABLE IF EXISTS %1s;', $schemas ) );
+			$settings = $wpdb->prefix . 'scsc_settings';
+			$wpdb->query( $wpdb->prepare( 'DROP TABLE IF EXISTS %1s;', $settings ) );
+		}
+
 		flush_rewrite_rules();
 	}
 }
